Combine axes into Series, Allow Axis to scale
From the design meeting, we decided to combine all chart elements into
the Series items, and do the majority of the configuration ns-3 side.

Change the Value vs Log axis to a setting on the axis, rather than
its own item.

Allow axes to define the behavior for when a point is plotted outside
their current range.

See: #21